+++ src/App.js
@@ -1,5 +1,5 @@
 import React, { useEffect } from "react";
-import Container from "./components/Container";
+import { Container } from "./components";
 import { keepTheme } from "./utils/theme";
 import { focusOnSearchBar, showSearchBarFocusHelp } from "./utils/searchBox";
