+++ src/components/Toggle/Toggle.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
 import PropTypes from "prop-types";
-import { ReactComponent as SunIcon } from "../assets/sun.svg";
-import { ReactComponent as MoonIcon } from "../assets/moon.svg";
-import { setTheme } from "../utils/theme";
+import { ReactComponent as SunIcon } from "../../assets/sun.svg";
+import { ReactComponent as MoonIcon } from "../../assets/moon.svg";
+import { setTheme } from "../../utils/theme";
 import "./Toggle.css";
 
 function Toggle({ parentCallback }) {
