+++ src/components/index.js
@@ -1,0 +1,11 @@
+import Container from "./Container/Container";
+import Footer from "./Footer/Footer";
+import Header from "./Header/Header";
+import Item from "./Item/Item";
+import ItemsSelected from "./ItemsSelected/ItemsSelected";
+import Results from "./Results/Results";
+import ScrollToTopButton from "./ScrollToTopButton/ScrollToTopButton";
+import Search from "./Search/Search";
+import Toggle from "./Toggle/Toggle";
+
+export { Container, Footer, Header, Item, ItemsSelected, Results, ScrollToTopButton, Search, Toggle }
